// open-tactics-core/src/domain/walk-effect.enum.ts
export enum WalkEffect {
    FLY = 'fly',
    FLOAT = 'float',
    TELEPORT = 'teleport',
    OVERJUMP = 'overjump',
    JUST_WEAK_JUMP = 'just-weak-jump',
    WATER_WALKING = 'walter-walking',
    NOT_AFFECT_WHILE_SWINNING = 'can-swin',
    CANNOT_SWIN = 'cannot-swin',
    WALK_ON_LAVA = 'walk-on-lava',
    IGNORE_SNOW_EFFECT = 'ignore-snow-effect'
}