// open-tactics-core/src/domain/specie-type.enum.ts
export enum SpecieType {

    /**
     * Sentient species
     */
    HUMAN = 'human',
    LIZARD_MAN = 'lizard-man',
    EAGLE_MAN = 'eagle-man',

    /**
     * Beast types
     */
    FIRE_DRAGON = 'fire-dragon',
    POISON_DRAGON = 'poison-dragon',
    COCKTRICE = 'cocktrice',
    OCTOPUS = 'octopus',
    GOLEM = 'golem',
    GRYPHON = 'gryphon',

     /**
      * Undead types
      */
    SKELETON = 'skeleton',
    GHOST = 'ghost',
    LICH = 'lich',
    ANGEL = 'angel'
}